backend/requirements.txt
Flask==3.1.0
flask-cors==5.0.0
Flask-Caching==2.3.0
requests==2.32.3
python-dotenv==1.0.1
anthropic==0.45.0
gunicorn==21.2.0
sgdata-sdk==0.2.1
redis==5.2.1
